% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{avgGen}
\alias{avgGen}
\title{Performance data generation according to an average approach. Linear model.}
\usage{
avgGen(its, sts, rank, q = NULL, skills = NULL, mean = 0, deviation = 1)
}
\arguments{
\item{its}{Number of items.}

\item{sts}{Number of students.}

\item{rank}{Number of skills the set of items has.}

\item{q}{Q-Matrix representing the skills required to succeed an item. skills per item matrix.{0,1} values. NULL by default.}

\item{skills}{Matrix representing the probability of each student to answer correctly to an item involving a specific skill. students per skills matrix.[0-1] values. NULL by default.}

\item{mean}{If the skills matrix is not provided to the function, this parameter is the mean used to generate that matrix according to the standard distribution. 0 by default.}

\item{deviation}{If the skills matrix is not provided to the function, this parameter is the standard deviation used to generate that matrix according to the standard distribution. 1 by default.}
}
\value{
A list with the following information:
 \item{results}{Students per items matrix where each element represents if the student answered correctly (1) or not (0) to the correspondent item.}
 \item{skills.matrix}{A student per skill matrix where each element represents the probability of a student to answer correctly to an item that involves the correspondent skill.}
 \item{q.matrix}{The Q-matrix used.}
}
\description{
This function generates students' performance data according to a Q-matrix sampling model where more than one skill can be required to succeed an item. \cr
Given a set of skills involved in the set of items, there is a matrix where each element represents the probability of a student to suceed an item which involves the correspondent skill. It's a students per skills matrix.[0-1] values.  \cr
According to the information provided by the Q-Matrix, if an item requires 2 skills, the probability of a student to answer correctly to that item will be the root mean square of the probabilities of this student to succeed items involving both skills.
}
\examples{
#####EXAMPLE 1 : Generation introducing the minimum number of parameters required #####

 # We generate student performance data about 3 students answering to 5 items and 3 skills.
 result <- avgGen(its = 5, sts = 3, rank = 3)

 # We extract the correspondent information

 performance  <- result$results
 performance
       [,1] [,2] [,3] [,4] [,5]
 [1,]    0    1    0    0    1
 [2,]    1    0    0    0    0
 [3,]    1    0    1    1    1

 skills  <- result$skills.matrix
 skills
         [,1]      [,2]      [,3]
 [1,] 0.5659460 0.7730670 0.1510989
 [2,] 0.2754485 0.3557638 0.5745654
 [3,] 0.0414596 0.1636330 0.9992421

 q.matrix  <- result$q.matrix
 q.matrix
       [,1] [,2] [,3] [,4] [,5]
 [1,]    1    1    0    1    0
 [2,]    1    1    0    1    1
 [3,]    1    0    1    1    1

#####EXAMPLE 2 : Generation modeling the mean and the standard deviation used #####

 # We generate student performance data about 3 students answering to 5 items and 3 skills.
 #For the stantard distribution (skills matrix), we use a mean of 0.5 and a standard deviation of 0.2.
 result <- minGen(its = 5, sts = 3, rank = 3,mean=0.5,deviation=0.2)

 # We extract the correspondent information

 performance  <- result$results
 performance
       [,1] [,2] [,3] [,4] [,5]
 [1,]    0    1    1    1    1
 [2,]    1    0    0    0    1
 [3,]    1    0    1    1    1

 skills  <- result$skills.matrix
 skills
         [,1]      [,2]      [,3]
 [1,] 0.6575190 0.6271332 0.7383802
 [2,] 0.7441959 0.5290334 0.6381248
 [3,] 0.6988804 0.5926346 0.6011667

 q.matrix  <- result$q.matrix
 q.matrix
       [,1] [,2] [,3] [,4] [,5]
 [1,]    0    0    1    1    1
 [2,]    1    1    0    1    0
 [3,]    1    1    1    1    0

 #####EXAMPLE 3 : Generation introducing a Q-matrix #####

 #We generate a Q-Matrix with 3 skills involved.
 q.matrix <- QgenInc(num.skills = 3,maxSkillsPerItem=2)
 q.matrix
       [,1] [,2] [,3] [,4] [,5] [,6]
 [1,]    1    0    0    1    1    0
 [2,]    0    1    0    1    0    1
 [3,]    0    0    1    0    1    1

# We generate student performance data about 3 students answering to 6 items, 3 skills and the Q Matrix we want.
# For the stantard distribution (skills matrix), we use a mean of 0.5 and a standard deviation of 0.2.
 result <- minGen(its = 6, sts = 3, rank = 3,q = q.matrix,mean=0.5,deviation=0.2)

 # We extract the correspondent information

 performance  <- result$results
 performance
       [,1] [,2] [,3] [,4] [,5] [,6]
 [1,]    1    1    0    1    0    1
 [2,]    0    0    0    0    1    0
 [3,]    1    0    1    0    0    0

 skills  <- result$skills.matrix
 skills
         [,1]      [,2]      [,3]
 [1,] 0.7293017 0.6951063 0.6243320
 [2,] 0.8131855 0.6306849 0.7204342
 [3,] 0.6953339 0.6937963 0.5450721

 q.matrix  <- result$q.matrix
 q.matrix
       [,1] [,2] [,3] [,4] [,5] [,6]
 [1,]    1    0    0    1    1    0
 [2,]    0    1    0    1    0    1
 [3,]    0    0    1    0    1    1

 #####EXAMPLE 4 : Generation introducing a Q-Matrix and a skills matrix #####

 #We generate a Q-Matrix with 3 skills involved.
 q.matrix <- QgenInc(num.skills = 3,maxSkillsPerItem=2)
 q.matrix
       [,1] [,2] [,3] [,4] [,5] [,6]
 [1,]    1    0    0    1    1    0
 [2,]    0    1    0    1    0    1
 [3,]    0    0    1    0    1    1

 #We generate a skills matrix  skills for 3 students, 3 skills, a mean of 0 and a standard deviation of 1.
 skills  <- skillsGen(sts = 3,mean = 0,deviation = 1,rank = 3)
 skills
         [,1]      [,2]      [,3]
 [1,] 0.003903748 0.9802114 0.5885409
 [2,] 0.900522227 0.3580975 0.4001344
 [3,] 0.978809077 0.9618280 0.1626563

 # We generate student performance data about 3 students answering to 6 items, 3 skills and the Q Matrixand skills
 matrix we want.
 result <- minGen(its = 6, sts = 3, rank = 3,q = q.matrix,skills = skills)

 # We extract the correspondent information
 performance  <- result$results
 performance
       [,1] [,2] [,3] [,4] [,5] [,6]
 [1,]    0    1    1    0    0    1
 [2,]    1    1    0    0    0    0
 [3,]    1    1    1    1    0    0

 skills  <- result$skills.matrix
 skills
         [,1]      [,2]      [,3]
 [1,] 0.003903748 0.9802114 0.5885409
 [2,] 0.900522227 0.3580975 0.4001344
 [3,] 0.978809077 0.9618280 0.1626563

 q.matrix  <- result$q.matrix
 q.matrix
       [,1] [,2] [,3] [,4] [,5] [,6]
 [1,]    1    0    0    1    1    0
 [2,]    0    1    0    1    0    1
 [3,]    0    0    1    0    1    1
}
\seealso{
\code{\link{QgenInc}} for the external generation of a Q-matrix.\cr
 \code{\link{QgenReg}} for the external generation of a Q-matrix.\cr
 \code{\link{extendQ}} for the external generation of a Q-matrix.\cr
 \code{\link{reduceQ}} for the external generation of a Q-matrix.\cr
 \code{\link{skillsGen}} for the external generation of the skills matrix.\cr
}

