% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{barComparison}
\alias{barComparison}
\title{Bar graph}
\usage{
barComparison(results, labels, col)
}
\arguments{
\item{results}{Vector with the values to be compared.}

\item{labels}{Vector with the names of the values to be plotted.}

\item{col}{Vector of colors for the bars.}
}
\description{
This function generates a bar graph to compare the performance of different approaches.
}
\examples{
x <- randGen(its = 5, sts = 3)$result
 x
       [,1] [,2] [,3] [,4] [,5]
 [1,]    0    1    1    0    1
 [2,]    0    1    1    1    0
 [3,]    0    1    0    1    1

 y <- randGen(its = 5, sts = 3)$result
 y
       [,1] [,2] [,3] [,4] [,5]
 [1,]    0    0    1    1    1
 [2,]    1    0    0    1    1
 [3,]    1    0    0    0    0

 z <- randGen(its = 5, sts = 3)$result
 z
       [,1] [,2] [,3] [,4] [,5]
 [1,]    1    1    0    0    1
 [2,]    1    0    0    0    0
 [3,]    0    1    1    1    1

 result1  <- compare (x,y)
 result2  <- compare(x,z)
 results  <- c(result1,result2)
 labels <- c('y','z')
 col <- c('red','green')
 barComparison(results,labels,col)
}

