% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{compare}
\alias{compare}
\title{Percentage of similarity between two \{0,1\} matrices.}
\usage{
compare(reference, result)
}
\arguments{
\item{reference}{Matrix used as reference.}

\item{result}{Matrix to analyze its performance or its similarity to the reference matrix.}
}
\value{
The percentage of similarity comparing element by element
}
\description{
This function returns the performance reached by a \{0,1\} matrix, comparing it to a reference (expected) matrix.
}
\examples{
x <- randGen(its = 5, sts = 3)$result
 x
       [,1] [,2] [,3] [,4] [,5]
 [1,]    0    1    1    0    1
 [2,]    0    1    1    1    0
 [3,]    0    1    0    1    1

 y <- randGen(its = 5, sts = 3)$result
 y
       [,1] [,2] [,3] [,4] [,5]
 [1,]    1    0    1    1    1
 [2,]    1    0    0    0    0
 [3,]    0    0    1    0    1

 similarity <- compare(reference = x, result = y)
 similarity
 [1] 80
}

