% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{extendQ}
\alias{extendQ}
\title{Q-matrix extension.}
\usage{
extendQ(q, extraItems)
}
\arguments{
\item{q}{Q-matrix to extend.}

\item{extraItems}{Number of items(columns) to add to the Q-Matrix}
}
\value{
\item{q}{Q matrix with the information about the skills required to succeed each item. Skills per items matrix.}
}
\description{
This function increases the number of items of a Q-Matrix randomly repeating some of its columns. It adds the number of items (columns) introduced as  a parameter.
}
\examples{
#We generate a Q-matrix with 4 skills involved and 3 skills as maximum involved per item.
 q.matrix <- QgenInc(num.skills = 4, maxSkillsPerItem = 3)
 q.matrix
       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
 [1,]    1    0    0    0    1    1    1    0    0     0     1     1     1     0
 [2,]    0    1    0    0    1    0    0    1    1     0     1     1     0     1
 [3,]    0    0    1    0    0    1    0    1    0     1     1     0     1     1
 [4,]    0    0    0    1    0    0    1    0    1     1     0     1     1     1

 # We add 2 more items
 new.qmatrix <- extendQ(q = q.matrix, extraItems = 2)
 new.qmatrix
       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14] [,15] [,16]
 [1,]    1    0    0    0    1    1    1    0    0     0     1     1     1     0     1     1
 [2,]    0    1    0    0    1    0    0    1    1     0     1     1     0     1     0     1
 [3,]    0    0    1    0    0    1    0    1    0     1     1     0     1     1     0     0
 [4,]    0    0    0    1    0    0    1    0    1     1     0     1     1     1     0     1
}
\seealso{
\code{\link{QgenReg}} for the generation of a Q-Matrix .\cr
 \code{\link{QgenInc}} for the generation of a Q-Matrix .\cr
 \code{\link{reduceQ}} to modify the size of the a Q-matrix.\cr
}

