% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{oimage}
\alias{oimage}
\title{Grid graphic of a matrix.}
\usage{
oimage(data, sort = F, ...)
}
\arguments{
\item{data}{Matrix to be represented.}
}
\description{
This function represents a matrix under a grid graph. Values closer to 1 will be represented as cells with a color closer to white and values closer to 0 will be cells closer to red.
}
\examples{
x <- randGen(its = 5, sts = 3)$result
 x
       [,1] [,2] [,3] [,4] [,5]
 [1,]    0    1    1    0    1
 [2,]    0    1    1    1    0
 [3,]    0    1    0    1    1

 oimage(t(x),xlab='Students',ylab = 'Items')
}

