% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{poksMatrixGen}
\alias{poksMatrixGen}
\title{POKS matrix generation}
\usage{
poksMatrixGen(its, depNum, trees, indirect.dependencies = FALSE)
}
\arguments{
\item{its}{Number of items.}

\item{depNum}{The total number of dependencies in the poks matrix.}

\item{trees}{The number of independent trees in the poks matrix.}

\item{indirect.dependencies}{A boolean determining if the poks matrix contains indirect dependencies. FALSE by default}
}
\value{
\item{dependencies}{The poks matrix}
}
\description{
This function generates the matrix needed by the POKS model. It gives the Partial Orders Knowledge Structure (POKS).
Meaning that, if a student answers correctly to an item B, he'll get a higher chance to succeed items related to B but easier.
This can be modeled as \eqn{P(A|B) = p1}. It's a \{0,1\} matrix where a 1 in the position (i,j) means that suceeding the item i increases the possibilities of succeeding the item j.\cr
In order to generate this matrix, parameters as the total number of dependencies, the number of independent trees or the introduction of indirect dependencies can be modeled.
A tree is a partial order structure where succeeding an item, the root, increases the probability of answering correctly to the items that are under it with a direct or indirect dependance.
The indirect dependance happens when suceeding an item A increases the possibilities of succeeding item B, and item B increases the possibilities of item C. So, indirectly, item A increases the possibilities of item C.
}
\examples{
##### EXAMPLE 1 : POKS matrix without indirect dependencies #####
 #We generate a POKS matrix with 6 items, 2 trees and 4 dependencies.
 poks.matrix <- poksMatrixGen(its = 6, depNum = 4,trees = 2)
 poks.matrix
       [,1] [,2] [,3] [,4] [,5] [,6]
 [1,]    0    1    1    0    0    0
 [2,]    0    0    0    0    0    0
 [3,]    0    0    0    0    0    0
 [4,]    0    0    0    0    1    1
 [5,]    0    0    0    0    0    0
 [6,]    0    0    0    0    0    0

##### EXAMPLE 2 : POKS matrix wit indirect dependencies #####
 #We generate a POKS matrix with 8 items, 2 trees and 6 dependencies.
 poks.matrix <- poksMatrixGen(its = 8, depNum = 6,trees = 2,indirect.dependencies = TRUE)
 poks.matrix
       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
 [1,]    0    1    0    1    0    0    0    0
 [2,]    0    0    1    0    0    0    0    0
 [3,]    0    0    0    0    0    0    0    0
 [4,]    0    0    0    0    0    0    0    0
 [5,]    0    0    0    0    0    1    0    1
 [6,]    0    0    0    0    0    0    1    0
 [7,]    0    0    0    0    0    0    0    0
 [8,]    0    0    0    0    0    0    0    0
}
\references{
\url{http://cran.r-project.org/web/packages/partitions/index.html}
}
\seealso{
\code{\link{poksGen}} for the generation of data according to the POKS model\cr
}

