% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SD_G.R
\name{print.SD_G}
\alias{print.SD_G}
\title{Print SD_G objects}
\usage{
\method{print}{SD_G}(x, ...)
}
\arguments{
\item{x}{A SD_G object}

\item{...}{Optional parameters.}
}
\value{
The sdg ranks
}
\description{
S3 method to print SD_G objects.
}
