% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{QgenInc}
\alias{QgenInc}
\title{Q-matrix generation with a variable number of skills required per item.}
\usage{
QgenInc(num.skills, maxSkillsPerItem)
}
\arguments{
\item{num.skills}{Number of skills involved.}

\item{maxSkillsPerItem}{Maximum number of skills required to succeed an item. It represents the maximum difficulty of an item.}
}
\value{
\item{q}{Q matrix with the information about the skills required to succeed each item. Skills per items matrix.}
}
\description{
Q-matrix generation where the number of items depends on the maximum number of skills per item chosen.
If we define 3 skills involved in the whole set of items, and we want the most difficult item to require 2 skills, we will have a number of items corresponding to all the combinations of 1 skill per item and all the combinations of 2 skills per item.
In case we want to set the most difficult item to involve the 3 skills, our Q-matrix would include a last column corresponding to an item involving all the skills.
}
\examples{
#We generate a Q-matrix with 4 skills involved and maximum 3 skills involved per item.
 q.matrix <- QgenInc(num.skills = 4, maxSkillsPerItem = 3)
 q.matrix
       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
 [1,]    1    0    0    0    1    1    1    0    0     0     1     1     1     0
 [2,]    0    1    0    0    1    0    0    1    1     0     1     1     0     1
 [3,]    0    0    1    0    0    1    0    1    0     1     1     0     1     1
 [4,]    0    0    0    1    0    0    1    0    1     1     0     1     1     1
}
\seealso{
\code{\link{minGen}} for the external Q-Matrix that can be provided to this function .\cr
 \code{\link{avgGen}} for the external Q-Matrix that can be provided to this function.\cr
 \code{\link{dinGen}} for the external Q-Matrix that can be provided to this function.\cr
}

