% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{QgenReg}
\alias{QgenReg}
\title{Q-matrix generation with the same number of skills required per item.}
\usage{
QgenReg(num.skills, skillsPerItem)
}
\arguments{
\item{num.skills}{Number of skills involved.}

\item{skillsPerItem}{Number of skills per item.}
}
\value{
\item{q}{Q matrix with the information about the skills required to succeed each item. Skills per items matrix.}
}
\description{
Q-matrix with dimensions corresponding to all the combinations of items involving the number of skills per item defined.
In this function, all the items have the the same number of skills involved.
}
\examples{
#We generate a Q-matrix with 4 skills involved and 2 skills involved per item.
 q.matrix <- QgenReg(num.skills = 4, skillsPerItem = 2)
 q.matrix
       [,1] [,2] [,3] [,4] [,5] [,6]
 [1,]    1    1    1    0    0    0
 [2,]    1    0    0    1    1    0
 [3,]    0    1    0    1    0    1
 [4,]    0    0    1    0    1    1
}
\seealso{
\code{\link{minGen}} for the external Q-Matrix that can be provided to this function.\cr
 \code{\link{avgGen}} for the external Q-Matrix that can be provided to this function.\cr
 \code{\link{dinGen}} for the external Q-Matrix that can be provided to this function.\cr
}

