% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{reduceQ}
\alias{reduceQ}
\title{Q-matrix reduction.}
\usage{
reduceQ(q, numberOfItems)
}
\arguments{
\item{q}{Q-matrix to reduce in size.}

\item{numberOfOtems}{Number of items(columns) of the final Q-Matrix}
}
\value{
\item{q}{Q matrix with the information about the skills required to succeed each item. Skills per items matrix.}
}
\description{
This function reduces the size of a Q-Matrix randomly extracting the columns (number of items) needed.
}
\examples{
#We generate a Q-matrix with 4 skills involved and 3 skills as maximum involved per item.
 q.matrix <- QgenInc(num.skills = 4, maxSkillsPerItem = 3)
 q.matrix
       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
 [1,]    1    0    0    0    1    1    1    0    0     0     1     1     1     0
 [2,]    0    1    0    0    1    0    0    1    1     0     1     1     0     1
 [3,]    0    0    1    0    0    1    0    1    0     1     1     0     1     1
 [4,]    0    0    0    1    0    0    1    0    1     1     0     1     1     1

 # We generate a matrix of just 10 items
 new.qmatrix <- reduceQ(q = q.matrix, numberOfItems = 10)
 new.qmatrix
       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
 [1,]    1    0    0    1    1    1    0    0    1     1
 [2,]    0    1    1    0    1    0    1    0    0     0
 [3,]    1    0    0    1    1    0    1    1    1     1
 [4,]    0    1    1    1    0    1    0    1    1     1
}
\seealso{
\code{\link{QgenReg}} for the generation of a Q-Matrix .\cr
 \code{\link{QgenInc}} for the generation of a Q-Matrix .\cr
 \code{\link{extendQ}} to modify the size of the a Q-matrix.\cr
}

