% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SD_G.R
\name{SD_G}
\alias{SD_G}
\alias{SD.G}
\title{The SDG}
\usage{
SD_G(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{A wide-format data frame.}

\item{alpha}{The significance level.}

\item{...}{Optional parameters.}
}
\value{
A SD_G object.
}
\description{
The Statistical Difference based Grouping Test (SD-G) fixing the limitation of non-parametric Scott-Knott ESD
}
\examples{
data <- read.csv("/Users/hlp/R Project/SDG/example.csv", encoding="UTF-8")
sdg <- SD_G(data)
print(sdg)
plot(sdg)

}
