\name{SDG-package}
\alias{SDG-package}
\alias{SDG}
\docType{package}
\title{
Synthetic Data Generation for the Educational Data Mining.
SDG
}
\description{
The goal of this package is to generate synthetic data. The package is very flexible while introducing parameters and very adaptive depending on the information the user owns. The structure of the library is based on a collection of functions that generate students' performance data according to different models. Then, there are also functions to generate different types of matrices and functions to plot the information in a really useful and understandable way.

}
\details{
\tabular{ll}{
Package: \tab SDG\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-17\cr

}

}
\author{
Javier Asenjo Villamayor\cr
Supervisor: Michel C. Desmarais

Maintainer: Who to complain to <javier.asenjo.villamayor@gmail.com>

}
\references{
- Beheshti, B. and Desmarais, M.C. (2014). Assessing Model Fit With Synthetic vs. Real Data, Polytechnique Montreal.\cr\cr
- Beheshti, B. and Desmarais, M.C. (2014). Predictive performance of prevailing approaches to skills assessment techniques: Insights from real vs. synthetic data sets 7th Conference on Educational Data Data Mining (EDM 2014), London, England, p. 409-410.\cr\cr
-Desmarais, M.C. and Pelczer, I. (2010). On the Faithfullness of Simulated Student Performance Data. In Proceedings of Educational Data Mining 2010 (EDM2010). Pittsburg, PA, Jun. 11-13, p. 21-30. 
}


\keyword{ package }
\seealso{
CDM - package : \url{http://cran.r-project.org/web/packages/CDM/}\cr
Partitions - package : \url{http://cran.r-project.org/web/packages/partitions/index.html}

}
\examples{
library(SDG)
}
