% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{skillsGen}
\alias{skillsGen}
\title{Skills per student matrix generation according to the standard distribution.}
\usage{
skillsGen(sts, mean, deviation, rank)
}
\arguments{
\item{sts}{Number of students.}

\item{mean}{Mean of the standard distribution.}

\item{deviation}{Standard deviation of the standard distribution.
It represents a noise while assigning skills to each student.}

\item{rank}{Number of skills the set of items has.}
}
\value{
\item{skills}{The skills per student matrix}
}
\description{
This function generates a skills per students matrix where each value represents the probability of a student i to answer correctly to an item involving the skill j.
}
\examples{
#We generate a skills matrix for 3 students, involving 4 skills, with a mean of 0 and a standard deviation of 1.
 skills  <- skillsGen(sts = 3, mean = 0, deviation = 1, rank = 4)
 skills
         [,1]      [,2]       [,3]      [,4]
 [1,] 0.2379563 0.4103992 0.97868219 0.1400159
 [2,] 0.7146010 0.7078044 0.59517022 0.4639706
 [3,] 0.9454237 0.9641969 0.03825092 0.8487321
}
\seealso{
\code{\link{minGen}} for the generation of a skills matrix that can be provided to that function.\cr
 \code{\link{avgGen}} for the generation of a skills matrix that can be provided to that function.\cr
}

