% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{successProbGen}
\alias{successProbGen}
\title{Students per items matrix generation according to the standard distribution.}
\usage{
successProbGen(its, sts, mean, deviation)
}
\arguments{
\item{its}{Number of items.}

\item{sts}{Number of students.}

\item{mean}{Mean of the standard distribution.}

\item{deviation}{Standard deviation of the standard distribution.
It represents a noise while assigning skills to each student.}

\item{rank}{Number of skills the set of items has.}
}
\value{
\item{succes.prob}{The skills matrix}
}
\description{
This function generates a matrix with dimensions students*items where each value represents the probability of a student i to success an item j.
}
\examples{
#We generate a success probabilty matrix for 3 students and 5 items, with a mean of 0 and a standard deviation of 1.
 success.prob <- successProbGen(its = 5, sts = 3, mean = 0, deviation = 1)
 success.prob
         [,1]      [,2]      [,3]      [,4]      [,5]
 [1,] 0.3657776 0.1630787 0.5882404 0.9314930 0.1885536
 [2,] 0.8937646 0.6969860 0.9854258 0.5449783 0.7714228
 [3,] 0.9655305 0.5075924 0.1038589 0.4710194 0.7395491
}
\seealso{
\code{\link{randGen}} for the generation of a success probability matrix that can be provided to that function.\cr
}

